# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pokerbots Competition API
  version: "1.0.0"
  description: |
    Student-facing and admin API for the Berkeley Pokerbots competition.

    * Base path: `/api/v1`
    * Auth: `X-Api-Key: <team_api_key>` **or** `Authorization: Bearer <team_api_key>`.
      Admin routes accept `X-Admin-Key: <admin_key>` or the key of a team flagged `is_admin`.
    * Errors always use the envelope `{ "error": { "code", "message", "details" } }` with codes
      `unauthorized`, `forbidden`, `not_found`, `validation_error`, `conflict`, `rate_limited`,
      `payload_too_large`, `not_ready`, `internal_error`.
    * Bot protocol: see `docs/BOT_PROTOCOL.md`.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
    Bearer:
      type: http
      scheme: bearer
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    TeamMember:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string, nullable: true }
        student_id: { type: string, nullable: true }
...
    TournamentConfig:
      type: object
      description: Engine configuration (see SPEC.md). All fields optional when patching.
      properties:
        table_size: { type: integer, default: 9 }
        break_threshold: { type: integer, default: 7 }
        min_table_size: { type: integer, default: 6 }
        starting_stack: { type: integer, default: 1000 }
        blinds:
          type: object
          description: "{type: geometric, starting_small, starting_big, multiplier, ante_starts_at_level?, ante_fraction_of_bb?} or {type: levels, levels: [...]}"
        level_advance:
          type: object
          properties:
            max_level_duration_secs: { type: integer, nullable: true }
            hands_per_level: { type: integer, nullable: true, description: "advance when the average hands per table in the level reaches this" }
            avg_stack_thresholds: { type: array, items: { type: array, items: { type: integer }, minItems: 2, maxItems: 2 } }
        action_timeout_ms: { type: integer, default: 1000 }
        series_length: { type: integer, default: 100 }
        rng_seed: { type: integer }
        max_hands_per_table: { type: integer, nullable: true }
    RunProgress:
      type: object
      properties:
        tournaments_total: { type: integer }
        tournaments_completed: { type: integer }
        tournaments_failed_or_cancelled: { type: integer }
    Run:
      type: object
      properties:
        run_id: { type: string, example: run_9f8e7d6c5b4a }
        kind: { type: string, enum: [nightly, ondemand] }
        status: { type: string, enum: [queued, running, finalizing, completed, failed, cancelled] }
        config:
          type: object
          properties:
            series_length: { type: integer }
            table_size: { type: integer }
            seed: { type: integer }
            action_timeout_ms: { type: integer }
            starting_stack: { type: integer }
            record_hands: { type: boolean }
            tournament: { $ref: '#/components/schemas/TournamentConfig' }
        participant_count: { type: integer }
        created_by: { type: string, description: "team id, 'admin' or 'scheduler'" }
        created_at: { type: string, format: date-time }
        snapshot_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        error: { type: string, nullable: true }
        nightly_date: { type: string, nullable: true }
        progress: { $ref: '#/components/schemas/RunProgress' }
        participants:
          type: array
          description: Only on GET /runs/{run_id}
          items:
            type: object
            properties:
              team_id: { type: string }
              team_name: { type: string }
              submission_id: { type: string }
...
    PlatformSettings:
      type: object
      properties:
        tournament: { $ref: '#/components/schemas/TournamentConfig' }
        nightly_enabled: { type: boolean }
        nightly_time_utc: { type: string, example: "07:00" }
        nightly_series_length: { type: integer }
        validation:
          type: object
          properties:
            build_enabled: { type: boolean }
            build_timeout_secs: { type: integer }
            trial_enabled: { type: boolean }
            trial_hands: { type: integer }
            trial_max_substitution_rate: { type: number }
            allow_auto_activate: { type: boolean }
        autoscale:
          type: object
          properties:
            enabled: { type: boolean }
            min_workers: { type: integer }
            max_workers: { type: integer }
            worker_concurrency: { type: integer }
            max_bots_per_worker: { type: integer }
            nightly_deadline_hours: { type: number }
            ondemand_deadline_minutes: { type: number }
            est_secs_per_participant: { type: number }
            scale_down_cooldown_secs: { type: integer }
        ondemand_runs_per_team_per_day: { type: integer }
        ondemand_max_series_length: { type: integer }
        smoke_timeout_ms: { type: integer }
        max_upload_bytes: { type: integer }
        registration_open: { type: boolean }
        record_hands: { type: boolean }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Not allowed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Validation:
      description: Validation error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
security:
  - ApiKey: []
  - Bearer: []
paths:
  /auth/register:
    post:
      tags: [auth]
      security: []
      summary: Create a team and receive its API key (shown once)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [team_name]
              properties:
                team_name: { type: string, minLength: 2, maxLength: 40 }
                members: { type: array, items: { $ref: '#/components/schemas/TeamMember' } }
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                type: object
                properties:
                  team_id: { type: string }
                  team_name: { type: string }
                  api_key: { type: string, example: pb_live_... }
                  created_at: { type: string, format: date-time }
        "403": { $ref: '#/components/responses/Forbidden' }
        "409": { $ref: '#/components/responses/Conflict' }
        "422": { $ref: '#/components/responses/Validation' }
  /me:
    get:
      tags: [auth]
      summary: Team profile, active submission and rate-limit status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  team_id: { type: string }
                  team_name: { type: string }
                  api_key_prefix: { type: string }
                  is_admin: { type: boolean }
                  suspended: { type: boolean }
                  members: { type: array, items: { $ref: '#/components/schemas/TeamMember' } }
                  active_submission_id: { type: string, nullable: true }
                  rate_limits:
                    type: object
                    properties:
                      ondemand_runs_per_day: { type: integer }
                      ondemand_runs_remaining: { type: integer }
                      ondemand_max_series_length: { type: integer }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /teams/{team_id}/rotate-key:
    post:
      tags: [auth]
      summary: Rotate the team API key (own team, or admin)
      parameters: [{ name: team_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: New key
          content:
            application/json:
              schema: { type: object, properties: { team_id: { type: string }, api_key: { type: string } } }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /admin/config:
    get:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Current platform settings
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PlatformSettings' } } } }
    patch:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Deep-merge a partial settings object
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PlatformSettings' } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/PlatformSettings' } } } }
        "422": { $ref: '#/components/responses/Validation' }
  /admin/teams:
    get:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: List teams
      responses:
        "200": { description: OK }
    post:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Provision a team (works even when registration is closed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [team_name]
              properties:
                team_name: { type: string }
                members: { type: array, items: { $ref: '#/components/schemas/TeamMember' } }
                is_admin: { type: boolean, default: false }
      responses:
        "201": { description: Created (includes api_key) }
        "409": { $ref: '#/components/responses/Conflict' }
  /admin/teams/{team_id}/suspend:
    post:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Suspend (or reinstate) a team; suspended teams are excluded from runs
      parameters: [{ name: team_id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { suspended: { type: boolean, default: true } } }
      responses:
        "200": { description: OK }
        "404": { $ref: '#/components/responses/NotFound' }
  /admin/workers:
    get:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Worker heartbeats and queue depth
      responses:
        "200": { description: OK }
  /admin/revalidate/{submission_id}:
    post:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Re-run the smoke test for a submission
      parameters: [{ name: submission_id, in: path, required: true, schema: { type: string } }]
      responses:
        "202": { description: Queued }
        "404": { $ref: '#/components/responses/NotFound' }
  /admin/autoscale:
    get:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Current autoscaling plan (per-run demand, desired workers) and the last action taken
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings: { type: object, description: autoscale settings (editable via PATCH /admin/config) }
                  plan:
                    type: object
                    properties:
                      desired_workers: { type: integer }
                      slots_needed: { type: integer }
                      concurrent_bots_needed: { type: integer }
                      reason: { type: string }
                      runs: { type: array, items: { type: object } }
                  last_action: { type: object, nullable: true }
...
  /admin/runs/nightly:
    post:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Trigger a nightly-style series now (full active snapshot)
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { series_length: { type: integer }, seed: { type: integer } } }
      responses:
        "202": { description: Queued, content: { application/json: { schema: { $ref: '#/components/schemas/Run' } } } }
  /admin/runs/{run_id}/jobs:
    get:
      tags: [admin]
      security: [{ AdminKey: [] }, { ApiKey: [] }]
      summary: Jobs belonging to a run (debugging)
      parameters: [{ name: run_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
...